+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - fix: start session replay after remote config and feature flags are executed during first app start ([#261](https://github.com/PostHog/posthog-android/pull/261))
+- use /flags instead of /decide ([#245](https://github.com/PostHog/posthog-android/pull/245))
 
 ## 3.18.0 - 2025-06-12